backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
python-dotenv
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic